make Watch a final class
